Add translations to the RejectionReason
