+++ src/components/Notification/RejectionReason/interfaces.ts
@@ -1,0 +1,14 @@
+export interface ITranslations {
+  reason: string;
+  contactModerator: string;
+}
+
+export interface IContainerProps {
+  className?: string;
+  moderatorEmail: string;
+  message: string;
+}
+
+export interface IComponentProps extends IContainerProps {
+  translations?: ITranslations;
+}
